Simplify agent contract and reuse socket test helper
