+++ agent/src/main/kotlin/ru/souz/agent/Agent.kt
@@ -1,7 +1,9 @@
 package ru.souz.agent
 
 import kotlinx.coroutines.flow.Flow
+import ru.souz.agent.graph.StepInfo
 import ru.souz.agent.state.AgentContext
+import ru.souz.graph.Node
 import ru.souz.llms.LLMRequest
 import ru.souz.llms.LLMResponse
 
@@ -26,14 +28,27 @@ data class ActiveRunInput(
     val input: String,
 )
 
-interface Agent {
-    val sideEffects: Flow<AgentStreamChunk>
-
-    suspend fun execute(ctx: AgentContext<String>): String
-    suspend fun cancelActiveJob()
-}
-
 data class AgentExecutionResult(
     val output: String,
     val context: AgentContext<String>,
 )
+
+internal typealias GraphStepCallback =
+    (step: StepInfo, node: Node<Any?, Any?>, from: AgentContext<Any?>, to: AgentContext<Any?>) -> Unit
+
+/** Optional capability for publishing input into an open execution. */
+internal interface ActiveRunSteer {
+    suspend fun submitToActiveRun(build: suspend () -> ActiveRunInput?): Boolean
+}
+
+internal interface Agent {
+    val sideEffects: Flow<AgentStreamChunk>
+
+    suspend fun cancelActiveJob()
+
+    suspend fun execute(
+        ctx: AgentContext<String>,
+        onActiveRunReady: suspend () -> Unit = {},
+        onStep: GraphStepCallback? = null,
+    ): AgentExecutionResult
+}
