+++ agent/src/test/kotlin/ru/souz/SkillsGraphBasedAgentTest.kt
@@ -77,7 +77,7 @@ class SkillsGraphBasedAgentTest {
             nodesSummarization,
             nodesMemory,
         )
-        val result = skillsAgent.executeWithTrace(baseContext())
+        val result = skillsAgent.execute(baseContext())
 
         assertEquals("final", result.output)
         assertEquals(
@@ -95,7 +95,7 @@ class SkillsGraphBasedAgentTest {
             executed,
         )
 
-        skillsAgent.executeWithTrace(result.context.copy(input = "Again"))
+        skillsAgent.execute(result.context.copy(input = "Again"))
         assertEquals(PROVIDED_SYSTEM_PROMPT, result.context.systemPrompt)
     }
 
@@ -126,7 +126,7 @@ class SkillsGraphBasedAgentTest {
             nodesErrorHandling,
             nodesSummarization,
             nodesMemory,
-        ).executeWithTrace(baseContext())
+        ).execute(baseContext())
 
         assertEquals("friendly error", result.output)
         assertEquals(
